requirements.txt
streamlit
PyPDF2
pdf2image
pytesseract
requests
groq
python-dotenv
Pillow
gTTS
audio-recorder-streamlit
